polishing tokenization process
Previously, the ELAN2TEI script produced a vertical as its primary output. This is now changed: We produce a tokenized copy of the full document to 103_tei_w which should be copied to 010_manannot and annotated there by curators. Moreover, whenever executed, the script creates now a text vertical which takes the annotations found in 010_manannot and integrates them into the text verticals.